telegram_location_bot/requirements.txt
Flask
python-dotenv
pyTelegramBotAPI
SQLAlchemy
pyotp
requests
